Extract common code from service/grpcweb.ts
