+++ src/service/grpcweb.ts
@@ -1,160 +1,19 @@
-import {filePathToPseudoNamespace, normaliseFieldObjectName, replaceProtoSuffix, getPathToRoot} from "../util";
 import {ExportMap} from "../ExportMap";
 import {Printer} from "../Printer";
 import {CodePrinter} from "../CodePrinter";
-import {
-  FileDescriptorProto, MethodDescriptorProto,
-  ServiceDescriptorProto
-} from "google-protobuf/google/protobuf/descriptor_pb";
-import {WellKnownTypesMap} from "../WellKnown";
-import {getFieldType, MESSAGE_TYPE} from "../ts/FieldTypes";
+import {FileDescriptorProto} from "google-protobuf/google/protobuf/descriptor_pb";
 import {CodeGeneratorResponse} from "google-protobuf/google/protobuf/compiler/plugin_pb";
+import {createFile, RPCMethodDescriptor, RPCDescriptor, GrpcServiceDescriptor} from "./common";
 
 export function generateGrpcWebService(filename: string, descriptor: FileDescriptorProto, exportMap: ExportMap): CodeGeneratorResponse.File[] {
   return [
-    createFile(generateTypescriptDefinition(descriptor, exportMap), `${filename}_service.d.ts`),
+    createFile(generateTypeScriptDefinition(descriptor, exportMap), `${filename}_service.d.ts`),
     createFile(generateJavaScript(descriptor, exportMap), `${filename}_service.js`),
   ];
 }
 
-function createFile(output: string, filename: string): CodeGeneratorResponse.File {
-  const file = new CodeGeneratorResponse.File();
-  file.setName(filename);
-  file.setContent(output);
-  return file;
-}
-
-type CallingTypes = {
-  requestType: string
-  responseType: string
-};
-
-function getCallingTypes(method: MethodDescriptorProto, exportMap: ExportMap): CallingTypes {
-  return {
-    requestType: getFieldType(MESSAGE_TYPE, method.getInputType().slice(1), "", exportMap),
-    responseType: getFieldType(MESSAGE_TYPE, method.getOutputType().slice(1), "", exportMap),
-  };
-}
-
-function isUsed(fileDescriptor: FileDescriptorProto, pseudoNamespace: string, exportMap: ExportMap) {
-  return fileDescriptor.getServiceList().some(service => {
-    return service.getMethodList().some(method => {
-      const callingTypes = getCallingTypes(method, exportMap);
-      const namespacePackage = pseudoNamespace + ".";
-      return (
-        callingTypes.requestType.indexOf(namespacePackage) === 0 ||
-        callingTypes.responseType.indexOf(namespacePackage) === 0
-      );
-    });
-  });
-}
-
-type ImportDescriptor = {
-  readonly namespace: string
-  readonly path: string
-};
-
-type RPCMethodDescriptor = {
-  readonly nameAsPascalCase: string,
-  readonly nameAsCamelCase: string,
-  readonly functionName: string,
-  readonly serviceName: string,
-  readonly requestStream: boolean
-  readonly responseStream: boolean
-  readonly requestType: string
-  readonly responseType: string
-};
-
-class RPCDescriptor {
-  private readonly grpcService: GrpcWebServiceDescriptor;
-  private readonly protoService: ServiceDescriptorProto;
-  private readonly exportMap: ExportMap;
-
-  constructor(grpcService: GrpcWebServiceDescriptor, protoService: ServiceDescriptorProto, exportMap: ExportMap) {
-    this.grpcService = grpcService;
-    this.protoService = protoService;
-    this.exportMap = exportMap;
-  }
-  get name(): string {
-    return this.protoService.getName();
-  }
-
-  get qualifiedName(): string {
-    return (this.grpcService.packageName ? `${this.grpcService.packageName}.` : "") + this.name;
-  }
-
-  get methods(): RPCMethodDescriptor[] {
-    return this.protoService.getMethodList()
-      .map(method => {
-        const callingTypes = getCallingTypes(method, this.exportMap);
-        const nameAsCamelCase = method.getName()[0].toLowerCase() + method.getName().substr(1);
-        return {
-          nameAsPascalCase: method.getName(),
-          nameAsCamelCase,
-          functionName: normaliseFieldObjectName(nameAsCamelCase),
-          serviceName: this.name,
-          requestStream: method.getClientStreaming(),
-          responseStream: method.getServerStreaming(),
-          requestType: callingTypes.requestType,
-          responseType: callingTypes.responseType,
-        };
-      });
-  }
-}
-
-class GrpcWebServiceDescriptor {
-  private readonly fileDescriptor: FileDescriptorProto;
-  private readonly exportMap: ExportMap;
-  private readonly pathToRoot: string;
-
-  constructor(fileDescriptor: FileDescriptorProto, exportMap: ExportMap) {
-    this.fileDescriptor = fileDescriptor;
-    this.exportMap = exportMap;
-    this.pathToRoot = getPathToRoot(fileDescriptor.getName());
-  }
-
-  get filename(): string {
-    return this.fileDescriptor.getName();
-  }
-
-  get packageName(): string {
-    return this.fileDescriptor.getPackage();
-  }
-
-  get imports(): ImportDescriptor[] {
-    const dependencies = this.fileDescriptor.getDependencyList()
-      .filter(dependency => isUsed(this.fileDescriptor, filePathToPseudoNamespace(dependency), this.exportMap))
-      .map(dependency => {
-        const namespace = filePathToPseudoNamespace(dependency);
-        if (dependency in WellKnownTypesMap) {
-          return {
-            namespace,
-            path: WellKnownTypesMap[dependency],
-          };
-        } else {
-          return {
-            namespace,
-            path: `${this.pathToRoot}${replaceProtoSuffix(replaceProtoSuffix(dependency))}`
-          };
-        }
-      });
-    const hostProto = {
-      namespace: filePathToPseudoNamespace(this.filename),
-      path: `${this.pathToRoot}${replaceProtoSuffix(this.filename)}`,
-    };
-    return [ hostProto ].concat(dependencies);
-  }
-
-  get services(): RPCDescriptor[] {
-    return this.fileDescriptor.getServiceList()
-      .map(service => {
-        return new RPCDescriptor(this, service, this.exportMap);
-      });
-  }
-}
-
-function generateTypescriptDefinition(fileDescriptor: FileDescriptorProto, exportMap: ExportMap) {
-  const serviceDescriptor = new GrpcWebServiceDescriptor(fileDescriptor, exportMap);
+function generateTypeScriptDefinition(fileDescriptor: FileDescriptorProto, exportMap: ExportMap): string {
+  const serviceDescriptor = new GrpcServiceDescriptor(fileDescriptor, exportMap);
   const printer = new Printer(0);
 
   // Header.
@@ -241,8 +100,8 @@ function generateTypescriptDefinition(fileDescriptor: FileDescriptorProto, expor
   return printer.getOutput();
 }
 
-function generateJavaScript(fileDescriptor: FileDescriptorProto, exportMap: ExportMap) {
-  const serviceDescriptor = new GrpcWebServiceDescriptor(fileDescriptor, exportMap);
+function generateJavaScript(fileDescriptor: FileDescriptorProto, exportMap: ExportMap): string {
+  const serviceDescriptor = new GrpcServiceDescriptor(fileDescriptor, exportMap);
   const printer = new Printer(0);
 
   // Header.
